# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayCat API
  description: |
    Unified Payment & Subscription Management System.
    A RevenueCat-compatible API for managing in-app purchases across iOS, Android, and Web (Stripe).
  version: 1.0.0
  contact:
    name: PayCat Support
  license:
    name: MIT
servers:
  - url: https://your-worker.workers.dev
    description: Production
  - url: http://localhost:8787
    description: Local Development
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/integrations:
    get:
      tags: [Integrations]
      summary: List integrations
      responses:
        '200':
          description: List of integrations

    post:
      tags: [Integrations]
      summary: Create integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationCreate'
      responses:
        '201':
          description: Integration created
  /v1/integrations/{id}/test:
    post:
      tags: [Integrations]
      summary: Test integration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test result
  /v1/events:
    get:
      tags: [Integrations]
      summary: List custom events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: event_name
          in: query
          schema:
            type: string
        - name: app_user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of events

    post:
      tags: [Integrations]
      summary: Track custom event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_user_id, event_name]
              properties:
                app_user_id:
                  type: string
                event_name:
                  type: string
                properties:
                  type: object
                timestamp:
                  type: integer
      responses:
        '201':
          description: Event tracked
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for your app (pk_live_xxx or pk_test_xxx)
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin session token
  schemas:
...
    IntegrationCreate:
      type: object
      required: [type, name, config]
      properties:
        type:
          type: string
          enum: [amplitude, mixpanel, segment, firebase, braze, slack, appsflyer, adjust, webhook]
        name:
          type: string
        config:
          type: object
          description: Type-specific configuration
        events:
          type: array
          items:
            type: string
...